Remove YAML logic into yaml.py
