+++ src/uwtools/config/support.py
@@ -5,7 +5,6 @@ from importlib import import_module
 from typing import Dict, Type, Union
 
 import yaml
-from f90nml import Namelist  # type: ignore
 
 from uwtools.exceptions import UWConfigError
 from uwtools.logging import log
@@ -15,15 +14,6 @@ INCLUDE_TAG = "!INCLUDE"
 
 
 # Public functions
-
-
-def add_yaml_representers() -> None:
-    """
-    Add representers to the YAML dumper for custom types.
-    """
-    yaml.add_representer(UWYAMLConvert, UWYAMLConvert.represent)
-    yaml.add_representer(Namelist, _represent_namelist)
-    yaml.add_representer(OrderedDict, _represent_ordereddict)
 
 
 def depth(d: dict) -> int:
@@ -74,31 +64,6 @@ def log_and_error(msg: str) -> Exception:
     """
     log.error(msg)
     return UWConfigError(msg)
-
-
-# Private functions
-
-
-def _represent_namelist(dumper: yaml.Dumper, data: Namelist) -> yaml.nodes.MappingNode:
-    """
-    Convert an f90nml Namelist to an OrderedDict, then represent as a YAML mapping.
-
-    :param dumper: The YAML dumper.
-    :param data: The f90nml Namelist to serialize.
-    """
-    namelist_dict = data.todict()
-    return dumper.represent_mapping("tag:yaml.org,2002:map", namelist_dict)
-
-
-def _represent_ordereddict(dumper: yaml.Dumper, data: OrderedDict) -> yaml.nodes.MappingNode:
-    """
-    Recursrively convert an OrderedDict to a dict, then represent as a YAML mapping.
-
-    :param dumper: The YAML dumper.
-    :param data: The OrderedDict to serialize.
-    """
-
-    return dumper.represent_mapping("tag:yaml.org,2002:map", from_od(data))
 
 
 class UWYAMLTag:
