+++ src/uwtools/tests/config/test_support.py
@@ -6,10 +6,8 @@ Tests for uwtools.config.jinja2 module.
 import logging
 from collections import OrderedDict
 
-import f90nml  # type: ignore
 import pytest
 import yaml
-from f90nml import Namelist
 from pytest import fixture, raises
 
 from uwtools.config import support
@@ -22,14 +20,6 @@ from uwtools.exceptions import UWConfigError
 from uwtools.logging import log
 from uwtools.tests.support import logged
 from uwtools.utils.file import FORMAT
-
-
-def test_add_yaml_representers():
-    support.add_yaml_representers()
-    representers = yaml.Dumper.yaml_representers
-    assert support.UWYAMLConvert in representers
-    assert OrderedDict in representers
-    assert Namelist in representers
 
 
 @pytest.mark.parametrize(
@@ -71,20 +61,6 @@ def test_log_and_error(caplog):
         raise support.log_and_error(msg)
     assert msg in str(e.value)
     assert logged(caplog, msg)
-
-
-def test_represent_namelist():
-    support.add_yaml_representers()
-    namelist = f90nml.reads("&namelist\n key = value\n/\n")
-    assert yaml.dump(namelist, default_flow_style=True).strip() == "{namelist: {key: value}}"
-
-
-def test_represent_ordereddict():
-    support.add_yaml_representers()
-    ordereddict_values = OrderedDict([("example", OrderedDict([("key", "value")]))])
-    assert (
-        yaml.dump(ordereddict_values, default_flow_style=True).strip() == "{example: {key: value}}"
-    )
 
 
 class Test_UWYAMLConvert:
